Moving send page into tab
