+++ app/components/views/TransactionsPage/SendTab/index.js
@@ -1,53 +1,197 @@
-// @flow
-import React from "react";
-import PropTypes from "prop-types";
+import React, { Component } from "react";
 import { autobind } from "core-decorators";
-import TabContent from "../../../TabbedPage/TabContent";
+import { compose, not, eq, get } from "../../../../fp";
+import service from "../../../../connectors/service";
+import settings from "../../../../connectors/settings";
 import send from "../../../../connectors/send";
+import SendPage from "./Page";
+import ErrorScreen from "../../../ErrorScreen";
+import { restrictToStdDecimalNumber } from "../../../../helpers/strings";
+import { FormattedMessage as T } from "react-intl";
 
-const propTypes = {
-};
+const BASE_OUTPUT = { destination: "", amountStr: "" };
 
 @autobind
-class SendTab extends React.Component {
-
+class Send extends Component {
   constructor(props) {
     super(props);
+    this.state = this.getInitialState();
   }
 
+  getInitialState() {
+    return {
+      isShowingConfirm: false,
+      isSendAll: false,
+      hastAttemptedConstruct: false,
+      account: this.props.defaultSpendingAccount,
+      outputs: [{ key: 0, ...BASE_OUTPUT }]
+    };
+  }
+
+  componentWillMount() {
+    this.props.onClearConstructTxError();
+    this.props.onClearPublishTxError();
+    this.props.onClearSignTxError();
+    this.props.onClearPublishTxSuccess();
+  }
+  componentWillUnmount() {
+    this.onClearTransaction();
+  }
   render() {
-    return (
-      <TabContent>
-        <div className="tab-content-card">
-        send tab<br />
-        <br />
-        <br />
-        <br />
-        <br />
-        <br />
-        <br /><br />
-        <br />
-        <br /><br />
-        <br />
-        <br /><br />
-        <br />
-        <br /><br />
-        <br />
-        <br /><br />
-        <br />
-        <br /><br />
-        <br />
-        <br /><br />
-        <br />
-        <br /><br />
-        <br />
-        <br />
-        sdifjaosdf bla :)
-        </div></TabContent>
-    )
+    const {
+      onChangeAccount,
+      onAttemptSignTransaction,
+      onClearTransaction,
+      onShowConfirm,
+      onShowSendAll,
+      onHideSendAll,
+      onAttemptConstructTransaction,
+      onAddOutput,
+      getOnRemoveOutput,
+      getOnChangeOutputDestination,
+      getOnChangeOutputAmount,
+      getAddressError,
+      getAmountError
+    } = this;
+    const isValid = this.getIsValid();
+
+    return !this.props.walletService ? <ErrorScreen /> : (
+      <SendPage
+        {...{ ...this.props, ...this.state }}
+        {...{
+          isValid,
+          onChangeAccount,
+          onAttemptSignTransaction,
+          onClearTransaction,
+          onShowConfirm,
+          onShowSendAll,
+          onHideSendAll,
+          onAttemptConstructTransaction,
+          onAddOutput,
+          getOnRemoveOutput,
+          getOnChangeOutputDestination,
+          getOnChangeOutputAmount,
+          getAddressError,
+          getAmountError
+        }}
+      />
+    );
+  }
+
+  onChangeAccount(account) {
+    this.setState({ account });
+    this.onAttemptConstructTransaction();
+  }
+
+  onAttemptSignTransaction(privpass) {
+    const { unsignedTransaction, onAttemptSignTransaction } = this.props;
+    if (!privpass || !this.getIsValid()) return;
+    onAttemptSignTransaction && onAttemptSignTransaction(privpass, unsignedTransaction);
+    this.onClearTransaction();
+  }
+
+  onClearTransaction() {
+    this.setState(this.getInitialState(), this.props.onClearTransaction);
+  }
+  onShowSendAll() {
+    this.setState({ isSendAll: true }, this.onAttemptConstructTransaction);
+  }
+  onShowConfirm() {
+    if (!this.getIsValid()) return;
+    this.setState({ isShowingConfirm: true });
+  }
+
+  onAttemptConstructTransaction() {
+    const { onAttemptConstructTransaction, unitDivisor } = this.props;
+    const confirmations = 0;
+    if (this.getHasEmptyFields()) return;
+    this.setState({ hastAttemptedConstruct: true });
+    if (this.getIsInvalid()) return;
+
+    if (!this.getIsSendAll()) {
+      onAttemptConstructTransaction && onAttemptConstructTransaction(
+        this.state.account.value,
+        confirmations,
+        this.state.outputs.map(({ amountStr, destination }) => ({
+          destination,
+          amount: parseFloat(amountStr) * unitDivisor
+        }))
+      );
+    } else {
+      onAttemptConstructTransaction && onAttemptConstructTransaction(
+        this.state.account.value,
+        confirmations,
+        this.state.outputs.map(({ destination }) => ({
+          destination,
+        })),
+        true
+      );
+    }
+  }
+
+  onAddOutput() {
+    const { outputs } = this.state;
+    this.setState({ outputs: [...outputs, { key: outputs.length, ...BASE_OUTPUT }] });
+  }
+
+  getOnRemoveOutput(key) {
+    return () => this.setState(
+      { outputs: this.state.outputs.filter(compose(not(eq(key)), get("key"))) },
+      this.onAttemptConstructTransaction
+    );
+  }
+
+  getOnChangeOutputDestination(key) {
+    return destination => this.setState({
+      outputs: this.state.outputs.map(o => (o.key === key) ? { ...o, destination } : o)
+    });
+  }
+
+  getOnChangeOutputAmount(key) {
+    return amountStr => this.setState({
+      outputs: this.state.outputs.map(o => (o.key === key) ? {
+        ...o, amountStr: restrictToStdDecimalNumber(amountStr)
+      } : o)
+    });
+  }
+
+  getIsInvalid() {
+    return !!this.state.outputs.find((o, index) => (
+      this.getAddressError(index) || this.getAmountError(index)
+    ));
+  }
+
+  getHasEmptyFields() {
+    return !!this.state.outputs.find(({ destination, amountStr }) => !destination || (!amountStr && !this.state.isSendAll));
+  }
+
+  getIsValid() {
+    return !!(
+      !this.getIsInvalid() &&
+      this.props.unsignedTransaction &&
+      !this.props.isConstructingTransaction
+    );
+  }
+  getIsSendAll() {
+    return this.state.isSendAll;
+  }
+  getAddressError(key) {
+    // do some more helper address checking here
+    // possibly check for Ds/Dc Ts/Tc and length at the least
+    // later can do full address validtion from dcrutil code
+    // (actually we should do this in selectors once state is moved to redux)
+    const { outputs } = this.state;
+    const { destination } = outputs[key];
+    if (!destination) return <T id="send.errors.invalidAddress" m="*Please enter a valid address" />;
+  }
+
+  getAmountError(key) {
+    const { outputs, isSendAll} = this.state;
+    const { amountStr } = outputs[key];
+    const amount = parseFloat(amountStr);
+    if (isNaN(amount) && !isSendAll) return <T id="send.errors.invalidAmount" m="*Please enter a valid amount" /> ;
+    if (amount <= 0 && !isSendAll) return <T id="send.errors.negativeAmount" m="*Please enter a valid amount (> 0)" />;
   }
 }
 
-SendTab.propTypes = propTypes;
-
-export default send(SendTab);
+export default service(settings(send(Send)));
