+++ app/style/Layout.less
@@ -57,7 +57,7 @@
   position: absolute;
   left: 0px;
   right: 0px;
-  top: 176px;
+  top: 180px;
   bottom: 0px;
 }
 
@@ -70,4 +70,8 @@
 
 .tab-content-card {
   background-color: white;
+}
+
+.tab-content-blur {
+  filter: blur(6px);
 }
