+++ app/style/PassphraseModal.less
@@ -11,6 +11,13 @@
   background-color: rgba(12, 30, 62, 0.5);
 }
 
+.tab-content-wrapper > .passphrase-modal {
+  right: -60px;
+  left: -60px;
+  top: -60px;
+  height: 161%;
+}
+
 .passphrase-modal-header {
   min-height: 0px;
   padding-bottom: 10px;
